opensearch-kb/config.py
import os
from dotenv import load_dotenv

load_dotenv()

# OpenSearch 配置
OPENSEARCH_HOST = os.getenv("OPENSEARCH_HOST", "localhost")
OPENSEARCH_PORT = int(os.getenv("OPENSEARCH_PORT", "9200"))
OPENSEARCH_USER = os.getenv("OPENSEARCH_USER", "admin")
OPENSEARCH_PASSWORD = os.getenv("OPENSEARCH_PASSWORD", "admin")

# 索引配置
INDEX_NAME = "vector_search_index"
PIPELINE_NAME = "text_embedding_pipeline"

# 向量模型配置
MODEL_NAME = "sentence-transformers/all-MiniLM-L6-v2"
VECTOR_DIMENSION = 384  # 模型输出维度

# API 配置
API_HOST = os.getenv("API_HOST", "0.0.0.0")
API_PORT = int(os.getenv("API_PORT", "8000")) 