opensearch-kb/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opensearch-py==2.3.2
python-dotenv==1.0.0
pydantic==2.4.2
numpy==1.24.3
sentence-transformers==2.2.2 